requirements.txt
# Low-Light People Detection with YOLOv8
# Python 3.8+ required

# Core dependencies
ultralytics>=8.0.0           # YOLOv8 framework
opencv-python>=4.8.0         # Image processing and CLAHE
numpy>=1.24.0                # Array operations
Pillow>=10.0.0               # Image I/O
torch>=2.0.0                 # PyTorch (required by ultralytics)
torchvision>=0.15.0          # Vision utilities

# Visualization and monitoring
matplotlib>=3.7.0            # Plotting and visualization
seaborn>=0.12.0              # Statistical visualizations
tensorboard>=2.13.0          # Training monitoring
wandb>=0.15.0                # Weights & Biases (optional)

# Data handling
PyYAML>=6.0                  # YAML configuration files
tqdm>=4.65.0                 # Progress bars
pandas>=2.0.0                # Data analysis

# Development tools
jupyter>=1.0.0               # Notebooks for experimentation
ipython>=8.12.0              # Enhanced Python shell

# Optional: Data annotation
labelImg>=1.8.6            # Annotation tool
